\name{print.plaqreffect}
\alias{print.plaqreffect}
\title{Print a plaqreffect object}
\description{
Print an object generated by nonlinEffect.
}
\usage{
\method{print}{plaqreffect}(x, ...)
}
\arguments{
  \item{x}{ an object returned from nonlinEffect.}
  \item{...}{optional arguments.}
}
\author{Adam Maidman}